requirements.txt
aniso8601==8.0.0
beautifulsoup4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
folium==0.12.1
googlemaps==4.4.5
html5lib
icecream~=2.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
JPype1
jsonify==0.5
jupyter==1.0.0
konlpy==0.5.2
lxml
MarkupSafe==1.1.1
mccabe
mysql-connector
nltk
numpy~=1.22.3
pandas~=1.4.3
pandas-datareader==0.9.0
pep8==1.7.1
plotly==5.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
seaborn
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
tensorflow==2.7.0
tensorflow-gpu==2.7.0
tqdm
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
mysqlclient
djangorestframework

django~=4.0.6
imutils~=0.5.4
opencv-python~=4.6.0.66
matplotlib~=3.5.2
easyocr~=1.5.0
torch~=1.8.0
transformers~=4.20.1
fastai~=2.7.7
gluonnlp~=0.10.0
kobert~=0.2.3